+++ project_test/tests/003_fields.py
@@ -3,10 +3,28 @@ import os
 import pytest
 
 from crispy_forms_foundation.layout import (Layout, InlineField,
-                                            InlineSwitchField)
+                                            InlineSwitchField, FakeField)
 
 from project_test.tests.forms import BasicInputForm, BoolInputForm
 #from project_test.tests.utils import write_output
+
+
+def test_fakefield(output_test_path, render_output, rendered_template,
+                     helper, client):
+    form = BasicInputForm()
+    pack = helper.template_pack
+
+    helper.layout = Layout(
+        FakeField('simple')
+    )
+
+    rendered = rendered_template(form, helper=helper)
+
+    attempted = render_output(os.path.join(output_test_path, pack,
+                                           "test_fakefield.html"))
+    #write_output(output_test_path, pack, "test_fakefield.html", rendered)
+
+    assert attempted == rendered
 
 
 def test_inlinefield(output_test_path, render_output, rendered_template,